refactor: better checks and naming of variables
We use uppercase for all exported variables and lower case for the ones
used inside the script only.
Checks for required variables uses bash variable expansion

+++ stop_transmission.sh
@@ -6,7 +6,7 @@
 set -eE
 failure() {
   local lineno=$1; local msg=$2
-  echo "$(basename "$0"): failed at $lineno: $msg"
+  echo "$(basename "$0"): failed at ${lineno}: ${msg}"
 }
 trap 'failure ${LINENO} "$BASH_COMMAND"' ERR
 
@@ -16,43 +16,37 @@ if ! command -v jq &>/dev/null; then
   echo "jq could not be found"; echo "Please install jq"; exit 1
 fi
 
-# Check if the mandatory environment variables are set.
-if [[ -z $NETNS_NAME ]]; then
-  echo "$(basename "$0") script requires:"
-  echo "NETNS_NAME - The name of the namespace"
-  exit 1
-fi
-
-SETTINGS=/home/felipe/.config/transmission/settings.json
-DEFAULT_PORT="62021" # default port when not using VPN
+_debug=${DEBUG:-false}
+settings=/home/felipe/.config/transmission/settings.json
+default_port="62021" # default port when not using VPN
 
 # Checks that Transmission settings file exist
-if [[ ! -f $SETTINGS ]]; then
-  >&2 echo "Cannot modify Transmission settings. $SETTINGS not found"; exit 1
+if [[ ! -f ${settings} ]]; then
+  >&2 echo "Cannot modify Transmission settings. ${settings} not found"; exit 1
 fi
 
 # Check if running as root/sudo
-[ ${EUID:-$(id -u)} -eq 0 ] || exec sudo -E "$(readlink -f "$0")" "$@"
+[ "${EUID:-$(id -u)}" -eq 0 ] || exec sudo -E "$(readlink -f "$0")" "$@"
 
 ##########################################
 # Stop transmission-daemon if it's active
 if pidof transmission-daemon > /dev/null; then
-  if [[ $DEBUG == true ]]; then echo "transmission-daemon is active. Stopping..."; fi
+  if [[ ${_debug} == true ]]; then echo "transmission-daemon is active. Stopping..."; fi
   # Kill all process inside the namespace or at least just transmission
-  ip netns pids "$NETNS_NAME" | xargs kill -9 > /dev/null 2>&1 || kill -9 "$(pidof transmission-daemon)"
+  ip netns pids "${NETNS_NAME:?}" | xargs kill -9 > /dev/null 2>&1 || kill -9 "$(pidof transmission-daemon)"
 fi
 
 # Changes port back to its default
-PORT=$(jq -r '."peer-port"' $SETTINGS) # retrieve port from settings file
-if [[ $DEBUG == true ]]; then echo "VPN port inside Transmission settings file: $PORT"; fi
+port=$(jq -r '."peer-port"' "${settings}") # retrieve port from settings file
+if [[ ${_debug} == true ]]; then echo "VPN port inside Transmission settings file: ${port}"; fi
 
 # Check if port needs modifying
-if (( PORT == DEFAULT_PORT )); then
-  if [[ $DEBUG == true ]]; then echo "Port rule is already at default value"; fi
+if (( port == default_port )); then
+  if [[ ${_debug} == true ]]; then echo "Port rule is already at default value"; fi
   exit 0
 fi
 
-tempSettings=$(jq '."peer-port"'="$DEFAULT_PORT" $SETTINGS)
-printf "%s" "$tempSettings" > $SETTINGS
+tempSettings=$(jq '."peer-port"'="${default_port}" "${settings}")
+printf "%s" "${tempSettings}" > "${settings}"
 
-if [[ $DEBUG == true ]]; then echo "Port rule back to default value"; fi
+if [[ ${_debug} == true ]]; then echo "Port rule back to default value"; fi
